+++ GHEtool/VariableClasses/Borehole.py
@@ -293,8 +293,9 @@ class Borehole(BaseClass):
             # initiate pipe
             pipe = self.pipe_data.pipe_model(k_s if isinstance(k_s, (float, int)) else k_s(depth, D), borehole)
 
-            return pipe.effective_borehole_thermal_resistance(self.flow_data.mfr(fluid_data=self.fluid_data, **kwargs),
-                                                              self.fluid_data.cp(**kwargs))
+            return pipe.effective_borehole_thermal_resistance(
+                self.flow_data.mfr_borehole(fluid_data=self.fluid_data, **kwargs),
+                self.fluid_data.cp(**kwargs))
 
         if 'temperature' in kwargs:
             kwargs_new = copy.deepcopy(kwargs)