+++ GHEtool/VariableClasses/FlowData/_FlowData.py
@@ -6,9 +6,9 @@ from GHEtool.VariableClasses.BaseClass import BaseClass
 class _FlowData(ABC, BaseClass):
 
     @abc.abstractmethod
-    def vfr(self, **kwargs):
+    def vfr_borefield(self, **kwargs):
         """
-        This function returns the volume flow rate [l/s].
+        This function returns the volume flow rate for the entire borefield [l/s].
 
         Parameters
         ----------
@@ -20,9 +20,37 @@ class _FlowData(ABC, BaseClass):
         """
 
     @abc.abstractmethod
-    def mfr(self, **kwargs):
+    def mfr_borefield(self, **kwargs):
         """
-        This function returns the mass flow rate [kg/s].
+        This function returns the mass flow rate for the entire borefield [kg/s].
+
+        Parameters
+        ----------
+        kwargs
+
+        Returns
+        -------
+
+        """
+
+    @abc.abstractmethod
+    def vfr_borehole(self, **kwargs):
+        """
+        This function returns the volume flow rate for a single borehole [l/s].
+
+        Parameters
+        ----------
+        kwargs
+
+        Returns
+        -------
+
+        """
+
+    @abc.abstractmethod
+    def mfr_borehole(self, **kwargs):
+        """
+        This function returns the mass flow rate for a single borehole [kg/s].
 
         Parameters
         ----------